+start-stop button pre jednotlive pece; +osetrenia chybovych stavov; +adresy pre t1-t2

+++ MOPROMAN (2023.10.03)/CSClient/record.cs
@@ -27,10 +27,10 @@ namespace nsAspur
         public Nullable<float> sobert_vykon { get; set; }
         public Nullable<float> tlak { get; set; }
         public string zmena { get; set; }
-        public Nullable<double> frekvencia { get; set; }
-        public Nullable<double> teplota_p1 { get; set; }
-        public Nullable<double> teplota_p2 { get; set; }
-        public Nullable<double> teplota_okruh { get; set; }
-        public Nullable<double> prietok_vody { get; set; }
+        public Nullable<float> frekvencia { get; set; }
+        public Nullable<float> teplota_p1 { get; set; }
+        public Nullable<float> teplota_p2 { get; set; }
+        public Nullable<float> teplota_okruh { get; set; }
+        public Nullable<float> prietok_vody { get; set; }
     }
 }
